requirements.txt
asgiref==3.8.1
Django==5.0.4
django-formtools==2.5.1
sqlparse==0.5.0
typing_extensions==4.11.0
